pyproject.toml
[project]
name = "model-deployer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.1",
    "django==6.0.3",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "jsonschema>=4.26.0",
    "pip==24.0",
    "psycopg2-binary==2.9.11",
    "python-dotenv==1.2.2",
    "pyyaml==6.0.3",
    "scikit-learn>=1.8.0",
    "sqlparse==0.5.5",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
]
